add fixes for problems during plug a thon
