+++ pat/consumer_tests/step_2.py
@@ -4,7 +4,9 @@ import logging
 import time
 from typing import TYPE_CHECKING
 
+from pat import consumer_mdib
 from sdc11073.consumer.consumerimpl import SdcConsumer
+from sdc11073.mdib import ConsumerMdib
 
 if TYPE_CHECKING:
     from sdc11073.consumer import ConsumerSubscription
@@ -40,11 +42,13 @@ def test_2a(consumer: SdcConsumer) -> bool:
     return False
 
 
-def test_2b(consumer: SdcConsumer) -> bool:
+def test_2b(consumer: SdcConsumer) -> tuple[bool, ConsumerMdib]:
     """The Reference Consumer renews at least one subscription once during the test phase; the Reference Provider grants subscriptions of at most 15 seconds (this allows for the Reference Consumer to verify if auto-renew works)."""  # noqa: E501, W505
     step = f'{__STEP__}b'
     max_subscription_duration = 15
     consumer.start_all()  # makes a subscription and requests 60 seconds per default in do_subscribe
+    # initializing the mdib immediately after subscribing avoids processing many reports with outdated mdib versions
+    mdib = consumer_mdib.get_mdib(consumer)
     for filter_text, subscription in consumer.subscription_mgr.subscriptions.items():
         filter_text: str
         subscription: ConsumerSubscription
@@ -64,7 +68,7 @@ def test_2b(consumer: SdcConsumer) -> bool:
             )
             # no need to continue if one subscription failed, otherwise test can wait indefinitely
             #   if granted expiration is very long
-            return False
+            return False, mdib
     subscriptions: list[ConsumerSubscription] = list(consumer.subscription_mgr.subscriptions.values())
     subscriptions.sort(key=lambda s: s.granted_expires)
     timeout = subscriptions[0].granted_expires + 1
@@ -73,6 +77,6 @@ def test_2b(consumer: SdcConsumer) -> bool:
     time.sleep(timeout)
     if subscriptions[0].is_subscribed and expires_at_before_renew < subscriptions[0].expires_at:
         logger.info('At least one subscription was auto-renewed successfully.', extra={'step': step})
-        return True
+        return True, mdib
     logger.error('No subscription was auto-renewed successfully.', extra={'step': step})
-    return False
+    return False, mdib
